Start recurrent jobs immediately
In a first version recurrent jobs were launched when the nextRun was
calculated. This meant that if a hourly job is created it won't get
executed until the hour has passed.

Usually you want to run the job immediately and then enqueue the next
execution at the required time. This commit fixes this. If is there a
need to only enqueue the job and not execute it the first time I'll add
the option.