Add isRunning() function and var on Job
IsRunning will allow external programs to know when a particular job
is running, which may be useful for say graceful shutdown.

Example Situation: Running scheduler from a console application and
when the console application is terminated, say using SIGTERM, it will
quit all of the Jobs using the Quit channel, but also try and wait for
currently running jobs to finish processing.